git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@198539 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: 1f86c1303755ea1806fd3e0cab53759cc1a5c6d8

+++ Online/Online/Gaucho/src/RPCRec.cpp
@@ -47,7 +47,7 @@ int RPCRec::analyseReply()
       {
         int namoff;
         char *nam;
-        if ((p1->type == (int)H_1DIM) || (p1->type == (int)H_PROFILE))
+        if ((p1->type == H_1DIM) || (p1->type == H_PROFILE))
         {
           namoff = p1->nameoff;//sizeof(DimHistbuff1);
         }
